deteksi suara hewan/requirements.txt
tkinter
Pillow
pygame
librosa
numpy
pandas
scikit-learn
sounddevice
soundfile
ttkbootstrap
joblib
